backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyodbc==5.1.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.35
authlib==1.3.2
httpx==0.27.2
itsdangerous==2.2.0
starlette[full]==0.38.6
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
pyyaml==6.0.2
aiosqlite==0.20.0
